+++ client/src/messages/gimbal.py
@@ -4,9 +4,9 @@ from messages.message import Message
 class Gimbal(Message):
     TYPE: str = 'gimbal'
 
-    def __init__(self, horizontal: int, vertical: int, center: bool = False, **kwargs):
-        self.horizontal: int = horizontal
-        self.vertical: int = vertical
+    def __init__(self, horizontal: float, vertical: float, center: bool = False, **kwargs):
+        self.horizontal: float = horizontal
+        self.vertical: float = vertical
         self.center: bool = center
 
     def __repr__(self) -> str:
